Sum of diagonal elements of a matrix using array
In this C Program to find Sum of diagonal elements of a matrix example, We declared single Two dimensional arrays Multiplication of size of 10 * 10. The below statements ask the User to enter the matrix size.
Next, we used C Programming for loop to iterate every cell present in a[3][3] matrix. Conditions inside the for loops ((rows < i) and (columns < j)) will ensure the program compiler, not to exceed the Matrix limit. Otherwise, the matrix will overflow. The scanf statement inside the for loop will store the user entered values in every individual array element such as a[0][0], a[0][1], etc.
In the next line, We have one more for loop to find Sum of Diagonal Elements of a Matrix.
User inserted values for C Program to find Sum of Diagonal Elements of a Multi-Dimensional Array example are: a[3][3] = {{1, 2, 3}, { 4, 5, 6}, {7, 8, 9}}
Row First Iteration: for(rows = 0; rows < 3; 0++)
The condition (0 < 3) is True.
Sum  = Sum + a[rows][rows]
Sum  = Sum + a[0][0]  => 0 + 10 = 10

Row Second Iteration: for(rows = 1; rows < 3; 1++)
The condition (1 < 3) is True.
Sum  = Sum + a[1][1]
Sum  = 1 + 5 = 6

Row Second Iteration: for(rows = 2; rows < 3; 2++)
The condition (2 < 3) is True.
Sum  = Sum + a[2][2]
Sum  = 6 + 9 = 15